enhancement: allow to mark division as natural person - db schema update path fix

+++ lib/upgradedb/mysql.2021051300.php
@@ -23,17 +23,19 @@
 
 $this->BeginTrans();
 
-$this->Execute("DROP VIEW vdivisions");
-$this->Execute("ALTER TABLE divisions ADD COLUMN firstname varchar(128) DEFAULT NULL");
-$this->Execute("ALTER TABLE divisions ADD COLUMN lastname varchar(128) DEFAULT NULL");
-$this->Execute("
-    CREATE VIEW vdivisions AS
-        SELECT d.*,
-            a.country_id as countryid, a.ccode, a.zip as zip, a.city as city, a.address,
-            (CASE WHEN d.firstname IS NOT NULL AND d.lastname IS NOT NULL THEN 1 ELSE 0 END) AS naturalperson
-        FROM divisions d
-            JOIN vaddresses a ON a.id = d.address_id;
-");
+if (!$this->ResourceExists('divisions.firstname', LMSDB::RESOURCE_TYPE_COLUMN)) {
+    $this->Execute("DROP VIEW vdivisions");
+    $this->Execute("ALTER TABLE divisions ADD COLUMN firstname varchar(128) DEFAULT NULL");
+    $this->Execute("ALTER TABLE divisions ADD COLUMN lastname varchar(128) DEFAULT NULL");
+    $this->Execute("
+        CREATE VIEW vdivisions AS
+            SELECT d.*,
+                a.country_id as countryid, a.ccode, a.zip as zip, a.city as city, a.address,
+                (CASE WHEN d.firstname IS NOT NULL AND d.lastname IS NOT NULL THEN 1 ELSE 0 END) AS naturalperson
+            FROM divisions d
+                JOIN vaddresses a ON a.id = d.address_id;
+    ");
+}
 
 $this->Execute("UPDATE dbinfo SET keyvalue = ? WHERE keytype = ?", array('2021051300', 'dbversion'));
